requirements.txt
aiohttp
pyjwt[crypto]